* commit '73a66ee7159dcde52083d1f80df4f9d4a8f56160':
  구글 제거
  0710 이환희 - 프렌차이즈 관리 구현, 유저 관리 구현, 사이드바 구현
  - 맵 폴리곤 클릭이벤트 확대기능 구현 완.
  0710 정상필 - 소셜로그인 닉네임 중복시 닉네임 재설정 기능 구현

+++ project/src/main/java/com/korea/project/controller/franchise/FranchiseController.java
@@ -51,7 +51,7 @@ public class FranchiseController {
         return "franchise/franchiseList";
     }
     @GetMapping("/franchiseboard")
-    public String showFranchiseDetail(@RequestParam("id") Long id, Model model) {
+    public String showFranchiseDetail(@RequestParam("id") int id, Model model) {
         FranchiseVO franchise = franchiseService.getFranchiseById(id);
         model.addAttribute("franchise", franchise);
         return "franchise/franchiseboard";
